kafka/requirements.txt
confluent-kafka==2.3.0
snowflake-connector-python==3.6.0
pydantic==2.6.0
python-dotenv==1.0.0
